less logging:
The verbosity of the ansible logging is causing the logs
to be unreadable. Turning down the verbosity for the general
public.  Also adding stdout_lines of the packstack install
and openstack-status.  Tested against rhel65, fedora20 and rhel7

Change-Id: I46ff33ddd31ee52121bfd260637cb8a2ccc2b480

add more debug at start of run

Change-Id: I178fc3df4b67e165617fd92111d82980449fbcee

logging updates

Change-Id: I530407ceb91352debf2af14d786e1a0ef0fa9add

updates for logging

Change-Id: Ib937f5f6bd5b7bacb57f0eeb866a2f029f773f04

updates for logging

Change-Id: I8ebb88f57f7640d8e8a758094133bd373af18600

updates for logging

Change-Id: Ib1175f59f8cd41518c930a7566ebfa73c4ab6059

revert changes to common-handlers

Change-Id: I895d9b8a428117915892cd641ca8613748797048