Tempest POC: add internal RedHat tempest setup
We need to compare the tempest results of our internal tool to the new tempest
configuration script, this commit makes it possible - but should be reverted
soon.

Added a separate 'tempest_rhosqe' role that duplicates, but simplifies a
lot of code that is in the 'tempest' role. This is because there were
problems merging it into the 'tempest' role - it conflicts with
setup-openstack-elements, the tempest configuration dictionary gets overwritten
in multiple configuration files (even in those outside of this repository) and
tempest_run is hard-coded to only use smoke tests. All of this could have been
fixed, but it would take time - this is a quick hack. It is just a POC and
will be removed soon anyway.

It uses the tag "tempest_rhosqe" instead of "tempest_setup,tempest_run" and the
variable "tempest_rhosqe" instead of "tempest", so that the values don't get
overwritten in configuration scripts.

It will run on the controller node and not on the extra tempest node, since
this is what the tools require.

Change-Id: I634f690c3e4a734a092b28cec1d405d6e86c4faf