ksgen: HACK: fix merging !lookup and another value
This hack fixes the issue that if foo: !lookup bar
merges with foo: baz it tries to merge LookupDirective and baz
and fails since LookupDirective has no 'extend'.

The current fix is to copy/overwrite foo: to baz is and is
acceptable in most of the cases but in the case of lists,
it should be extended instead of overwriting.

Change-Id: I840fdad6a35d08b828bddd25c8e80311fc8b13ad