Fix detection of version after last rhos-release changes
The latest version of rhos-release always enable the last puddle,
and in addition it enables poodle repository when a poodle is requested.
This means that multiple occurencies of the same package are possible,
this case is now handled with priority to poodles (if the poodle repository
is enabled, this is a poodle).

Change-Id: Ifd05cfae7bb03232839b4f8c9f463b398ab771a9