Fix back reference in the regexp; preserve spaces
A fix in ansible >= 1.7.2 broke the behavior of lineinfile with
regexp and backrefs in some conditions. More details here:
https://github.com/ansible/ansible-modules-core/issues/174
Khaleesi now requires ansible >= 1.7.2, escape the single quotes,
which restore the required behavior. Please note that this change
breaks lineinfile < 1.7.2.

Also, preserve the spaces at the beginning of the replaced line.

Change-Id: Ia2be03936630f3827fa35e05df0f90ee00d7b342