wait_for_ssh: fix hanging on stdin with pseudotty
usage of -tt with timeout wrapper causes ssh to hang
in same cases (when there is no tty during wait_for_ssh execution)
due to pseudo-tty messing with it, by enforcing < /dev/null
it seems to get fixed.

this also splits get-original boottime and do-reboot commands
to distinguish the case there is issue with rebooting itself
from the case when we start with un-sshable host.

Change-Id: Ie9b9e1d89797a0b341b7da8b6d7bba797f15104a