Support "--skip-file" in run-tempest.sh
Users can now add a black/white-list of tests according to run-tempest.sh
syntax (from run-tests.sh --help):

Format of skip file:
Whitelist is applied first. The blacklist is executed against the set
 of tests returned by the whitelist.
If whitelist is empty, all available tests are fed to blacklist.
If blacklist is empty, all tests from whitelist are returned.

The syntax for white-list and black-list is as follows:
- lines starting with # or empty are ignored
- lines starting with + are whitelisted
- lines starting with - are blacklisted
- lines not matching any of the above conditions are blacklisted

The match for each line gets added a ^ in the beginning,
so the regular expression should account for that.

For example, the following scenario:

    run all the smoke tests and scenario tests,
    but exclude the api.volume tests.

is implemented as:

    +.*smoke
    +tempest\.scenario
    -tempest\.api\.volume.*

Change-Id: I52cd44fe46bcc046f485be1a9d75e6375cf2056f