Alter configure-tempest.j2 template to set custom options in config_tempest.py call
I'm not sure if that's the best approach, however, with this change, we can for example
custom options being set in the config_tempest.py script.

Example:

in RHOS 5 that doesn't support ipv6, I can have something like this:

tester:
  tempest:
    config_options:
      - network_feature_enabled.ipv6 false

This will be set in the for loop in the configure-tempest.j2 and we will have it set in
tempest.conf, reducing the failures significantly.
In my tests I saw that mostly of the failures are because there is some misconfiguration
in the tempest.conf (like the ipv6 in the example).

This will not impact the actual jobs because it first check if the
tester.tempest.config_options is enabled, and only if it is enabled, it will enter in the
loop

Change-Id: I8e1def402f55130ec0d0629a07aeb44908b76390