Bump testtools version
Installation step "extra requirements" is overriding
the default tempest requirement from requirements.txt
(testtools>=0.9.36,!=1.2.0) with 0.9.34.
There are some Tempest (e.g. IPv6) tests using
3-parameter syntax of function AssertIn, which
is available only in testtools 0.9.36 and newer.
If is not necessary, this reference should be
remove or at least bumped
(this should be fixed in khaleesi-settings too).

Change-Id: I8eeb016b2564dede1be5d1ed54dff701b0dae68e