avoid strong dependency with the provider OS
If the provider OS is RHEL 7.1 but we try to deploy OSP on CentOS  7.0
or 7.1503, we will face a situation where ansible_distribution_version
and distro.version are not the same.

With this patch, will get the distribution verison directly from
distro.version, exactly like we do for the distro.name.

Change-Id: I2fe2682588b0ab73066cf91fc63773c736f66d7d