Run openstack undercloud install twice
The method for making changes to the undercloud's
configuration post-install, is to edit undercloud.conf,
and re-run `openstack undercloud install`. As such, we
should not allow in changes that break this behavior.

As an example, I caught this in manual testing here:
https://review.gerrithub.io/#/c/238249/3

The above patch passed the gate, but would have broke the
idempotency of the undercloud install.

* updated the logging for the undercloud install, ks change merged

Change-Id: I8a82f65573224cadc660c5546f80a0f397e51971