[ksgen] Fix default settings inheritance problem
This patch fix the default settings inheritance problem that occurred
when sub arguments with default settings in them have been provided.
The problem was that only defaults settings for the specific given
argument have been loaded instead of the default settings values of
all of its sub (inherited) arguments as well.

Change-Id: If81a9f1076f1d1de84afa63a55365e3b18fa671b