Updates scale command to pass scale values not update the plan
Previously, scaling up nodes of a certain type was done
by updating the tuskar plan and deploying the stack again
(update operation) without any scale parameters.

This change remove the plan update and passes the scaled
up parameters to the deploy command. In particular, when
updating an overcloud with more Ceph nodes, all
parameters need to be passed to the deploy command
that updates the stack.

Change-Id: I88e256c95b29d324b474e729a512f72d83d3a8ad