Avoid the download if image exist with get_url.
In virthost case, the stack user home is deleted so we have to keep the
image outside of this directory. And we can reuse the same image to put
it inside the undercloud. So for the undercloud part we just use the
get_url function to avoid the download if the pre-virt already put the
image here. If not (in case of BM) it will download the image itself.

Change-Id: Icf54e73eead6346f970bc08449098cbcb8ae86ae