Increase the wait_for used for booting VMs.
If there is a peak in VM boot requests -- for example when multijob
starts several jobs, each starting several VMs, the ammount of time
needed to get all in ACTIVE state can be really high.

My tests shows that with exponential backoff recently added, the time
needed for 20 VMs boot is about 25 minutes.

Change-Id: Icb8d7ab225b20eca4a30e6d31656d5272fdee3d1