Properly save default iptables rules on installation
Packstack does not touch existing iptables rules only if they
do not conflict with rules required by the services *and*
they are properly saved; so save them.

To be precise, it seems that packstack do not remove the existing rules
if they are in memory but there is no /etc/sysconfig/iptables file.
So remove also the old workaround which was creating an "empty"
/etc/sysconfig/iptables file, which (from testing) lead to
overwriting the existing rules.

Both changes combined makes sure that the existing rules are properly
saved after they are applied, making sure that packstack use them
as starting point. Also, the availability of the /etc/sysconfig/iptables
file make sure that the iptables service can really restart, in case
the old issue which led to the workaround is still valid.

Change-Id: Ic50f8ee00aefba7182d907fa2d26af750686702f