wait_for_ssh: warn when the path is not absolute
As ansible modules are being executed in different
pwd (in tmp) even on localhost, we do require
absolute path to the keyfile.

When the wait_for_ssh is going to fail due
to that keyfile is not valid file,
check if the 'absolute' path would be different,
and if so propose to the user that this may the reason.

As this seems to be quite common mistake.

Change-Id: If5ec3f6d94a34c510c6e333ac88638ea47f92a7a