RDO fix image building to use delorean correctly
If we use DIB_YUM_REPO_CONF instead of the delorean-repo
element, then yum-plugin-priorities does not get installed
in the overcloud images. This can lead to unexpected packages
being installed since we rely on yum priorities for delorean.

This change goes back to using the delorean-repo element for
image builds.

Change-Id: I051b9cc6ecfd6b97f9807322d38eb69bbe98aedc