Use a single delorean repo for all packages
Upstream tripleo is aimed at developers, specifically of
tripleo. So it makes sense for them to always want the
lastest tripleo packages. RDO-Manager on the other hand
is targeted more towards users, and stability is much
more important. As such, I think we should use a different
repo strategy for RDO-Manager than is used in tripleo.

This means we will also need to maintain a fork of the tripleo
docs, in order to document this simpler repo strategy.

Change-Id: I1b25b6070518235c1610eddd151d094d42f84c66