[roles][patch rpm] fix patching error handling
There should be error handling in patch_rpm
role, since some errors are legitimate patching
errors.

Real scenario example:
In any component repo (neutron, cinder...)
there is jenkins-config.yml file which not
included in the final package since it's
used for CI only. When changing this file
and trying to patch it with 'rdopkg mockbuild'
it will fail since there is no such file to
apply the patch on.

Change-Id: I64b366c34aad3a4887f19f0a55c3eeee8b2df481