Add opendaylight playbooks
OpenDaylight is one of neutron's backends.

This adds the necessary playbooks to configure opendaylight ( with L3 )
as neutron back-end and install opendaylight on dedicated node with one
of the following installation methods:

1. RPM (opendaylight)
2. URL of a zip file
3. Source (mvn build)

It also adds tempest tests configuration. Opendaylight isn't
supporting ipv6 and security group at the moment, so those are skipped.

This supports OpenStack all-in-one topology and not distributed.

Change-Id: I149537e3e82f0b17a60eee6b921ec4f96efa8376