Create and use undercloud.qcow2 'appliance'
More than half the time spent in our virtual rdo-manager jobs is spent
on creating the undercloud vm. We can save 20-40 minutes per job by
instead starting from a pre-built undercloud.qcow2. This can then
also be used by the RDO community for a "quickstart".

This change reorganizes all tasks related to both overcloud and undercloud
image creation and consumption for RDO-Manager into roles. It requires
that playbooks/installer/rdo-manager/roles is in the ansible roles path.

The standalone image building playbook was moved to the top-level directory
and instructions on how to use it are added to the README.

A playbook to fetch the images to the local workspace is added for CI use.

In order to use this in centos.ci, we need a different URL for jobs that
run on centos infra from jobs run outside. In order to do that, we need to
be able to override the image url for the centos infra jobs. This was not
possible with a '.' in the distro.full_version. For centos it does not make
sense to distinguish the minor release, as there is no such thing. Hence,
changed distro.full_version to 7 for centos. The settings/distro/centos-7.0.yml
file is left in place for compatibility, but should be removed once we make
all jobs use settings/distro/centos-7.yml instead.

Three new settings are added to rdo_manager/images in order to use this new
functionality:

- build_rdo_promote is meant to be used in the promote job. It will build
  all overcloud images and do the package install portion of installing the
  undercloud. Then it saves a snapshot of the undercloud storage for use
  in other CI jobs as well as an RDO-Manager quickstart.

- import_rdo_undercloud is meant to be used by all other rdo-manager virt jobs.
  It will download the undercloud image from the promote job and install it
  into libvirt.

- import_rdo_overcloud will replace the current import_rdo setting. This is
  just meant to be more explicit about what is being imported. The import_rdo
  setting is left in place until we can switch all jobs that use it to this
  setting or the import_rdo_undercloud setting.

Change-Id: I64ec42e9f0735d797ef111e6fd8501608841e0a7