jobs: add recheck support, upstream style messages
Now it's possible to retrigger the gate jobs with recheck
or rdo recheck. This comforms to upstream standards.

It will also output test results in a way that's familiar to upstream
developer, making out gating consistent.

Remove the silent: false which is the default anyway, and add a new
artifact to archive to support the new archive server feature.

Change-Id: I15f876012cff89dc02c3763fef8da362575085f9