Jenkins slave testing
So the idea is that I create a machine using the qeos provisioner
then I run a totally different ansible (not khaleesi) the internal slave
script there and then I verify that it worked with a basic suite of
tests (for now) but easily to be added in the future.

as a secondary case we can run the suite of tests on any slave to verify
that they are on the correct state.

Change-Id: I3afd51f42c52a4cc34b8a1bbca64549e50a4eeb6