[centosci] Add ability to avoid hufty
The hufty chassis only has 16Gb of RAM. This is not enough for
even the minimal RDO-Manager install and has been the root
cause of many failures in the periodic job. This patch adds
some logic to library/centosci.py to request a new node if
we get a hufty node.

Change-Id: I6aaa359a97dda8d53f997ff659e1c2d46ba8f28f