[RDOM] Lower default timeout for overcloud deploy
The overcloud deploy should not take more than an hour.
This patch lowers the default timeout for overcloud deploy
from 4 hours to 90 minutes, so that CI can fail faster when
it will fail due to timeout of overcloud deploy.

This would still be overridden if there is a timeout specified
in hw_env.deploy_timeout setting.

Change-Id: I237f835538e0528ce94ada49e5ccb1e97fb61169