Workaround: increase num_engine_workers bz1290949
The default heat.config value for num_engine_workers is set to the
number of CPU cores on the UC. This seems to be too low for an HA
deployment, since we are seeing timeouts on configuring the third
controller in a 3-controller deployment. This increases the default
value for num_engine_workers to 4.

Change-Id: I04e114228c799fff52ef69d15971bcd4336583e1