breakup the undercloud playbooks into logical steps
The goal w/ this patch is to start breaking down playbooks into independent
and idempotent steps as much as possible. The breakdown of the undercloud
should follow the tripleo docs as much as possible. Note, environment-setup,
undercloud etc, we're trying to reuse the same terms used in the docs.

The one step that can not be
idempotent is instack-virt-setup, until replaced by rdo a cleanup will have to
be executed before instack-virt-setup. Installing the undercloud, and
building/importing images should be idempotent and unique steps that can be
executed after instack-virt-setup (or other provisioner) has run.

In this review you will find that there is room for additional virt
provisioners, so that instack-virt-setup does *not* have to be the sole way
to stand up the virt nodes for the undercloud and overcloud.

The scope of this change should mainly be around reorganizing the playbooks,
removing uneeded code and updating playbooks to the best practices guidelines.

Baremetal testing
https://rhos-jenkins.rhev-ci-vms.eng.rdu2.redhat.com/view/POC/job/whayutin-bm-test-259645/

Depends-On: I58ca8d41b0b7d7b1e9b109f10e3dcfb09dc632a0@codeng
Change-Id: I66e31fccc2c9742844061a6e2c9b0bd6210ac200