Remove promote multi-job
RDO has created a rdo-infra repo[1] to house the promote
job jjb, in order to allow for other projects to
participate in gating delorean repo promotion without
having to go through khaleesi.

Khaleesi jobs will still be allowed in the testing phase
of the promote job, but will need to consume prebuilt
images (overcloud or undercloud). All jobs that run in
the testing phase will have to use and absolute timeout
of 2 hours.

I have also removed myself as an owner of the khaleesi
based jobs.

[1] https://github.com/redhat-openstack/rdo-infra

Change-Id: I3c41ffb99680639a2827036aaaf318208939350b