Adjust "small" and "medium" flavor on qeos7/rhos-jenkins
The small flavor "2" is too small on qeos7 for example for RHEL 7.2 base
images (which require a 10GB disk). So bump to "3", smaller value with
enough disk space.
At the same time, raise the requirements for medium to "4", which has
a bigger space, closer to the old configuration.

This could require a bigger change to flavor definitions (see CENTRALCI-1189).

Change-Id: Iafa43a5d7e1684e18d82f265bfc8dad890210482