Correctly set and read the credentials for test accounts
Do not hardcode the username and tenant for the non-admin testing
user/tenant.
Call the role for user creation for all integration tests (worst case:
no user is created). Also, at the same time, properly read the value
of admin_password without triggering an evaluation error (this fixes
an error on templating that went  unfortunately unnoticed).

Change-Id: Id4b540354006f3091c1b1caa5a79de90fe98d215