Support use of http-proxy-based installation
This allows the user to specify details of an http proxy server
to be used by the virthost and undercloud nodes when access to
some network resources is only available via the proxy.

For example if a remote install needs to use rhos-release to
access repos on the vpn, one can run squid proxy on a device
attached to the vpn and point their install at the proxy.

The user creates a file in the settings/installer/rdo_manager/proxy
(see default none.yml for needed values) and specifies it by setting
installer-proxy with ksgen.

Change-Id: I0673484fda4a92f9499e97a425d6a0adc9c384ad