Revert "Dynamic tester host"
This reverts commit d6169a5cf2c18e944fa869d4e60e932346345ff7.

The patch itself introduce some logical changes in the code (see
horizon integration tests, where actions to be executed on the controller
and on the tester have been conflated).
The lack of controller in some cases should be fixed by fixing the topology;
a lot of code assume the usage of controller to be able to execute
some operations on the controller, and this should not be changed, even
if the tester is going to be moved around.

Change-Id: Ib4f061feacec9a822f1005607ebeb11625a67570