Make the creation of the tester node explicit
This patch creates a new flag that explicitly state the need to
create a tester node and make the creation of the tester
really part of the topology.
The default is "yes" for now, until the jobs are fixed to explicitly
specify the flag. After that, the default can be changed to "no".

The new flag is --provisioner-create_tester=<yes/no>

Also, make rhosqe tester extend common, as other testers do.

Co-Authored-By: Luigi Toscano <ltoscano@redhat.com>

Change-Id: I5f5a80f5f5c51666fd619c7017cc233eeeb80ff7