Add rhos release 9 settings
This is required for running tests on OSP 9 release.

This is basically a copy of 8.0 release without workarounds
and styling changed from 2 spaces to 4 spaces.

In the future, we should work on better mechanism for adding
new releases and work on better version structure (at the
moment it looks like there is a lot of duplication).

Change-Id: I497d755fe0cc72c5c57a014102f50a13e86edb49