[rdo-manager] Run neutron-sanity-check
neutron-sanity-check is a helper script was added to neutron
to facilitate testing whether or not a running system supports all of
the Neutron configuration options that have been enabled.

Since every rdo-manager/OSP-d dpeloyment includes Neutron, it will
be reasonable (and effective) to run it every time to verify neutron
installed and configured properly.

This will run neutron-sanity-check against undercloud and overcloud
controller.

Change-Id: I88a0cf770697c64981230183a127a926f49bbc75